Fix major bug in dataBackup.sh
The script previously used all available older backup dirs as
--link-dest arguments to rsync, but rsync only accepts a maximum
of 20. So that caused the script to not actually do the backups since
it failed on this. Also added 'set -e' to the beginning of the
script so that it should fail on first failed command and return
a non-zero exit code as a whole when appropriate, previously it
always returned 0, independent on the actual error code. This change
should ensure that errors are propagated correctly.