Remove set -e in dataBackup.sh
It made the script fail for any small error, for instance if there was
one file in a directory that could not be backed up, it would case
the rest of the backup to not complete. The rationale for adding set -e
in the first place was supposedly to avoid a potential infinte loop
when deleting old dirs. Now I do that with " .. || exit 1" instead.
NOTE: I should probably fix the bug where the entire backup history can
be deleted if a backup is done after a long break, such as what just
happened...