Re-add vundle setup stuff to setup.sh
It seems that I need those lines since the git repos in ./vim/bundle
folder is not added to the project - and for some reason I cannot add
them to git the normal way. But I cannot be arsed to use submodules
either because it seems that I then would need to know the git url:s
which I suppose would be ok in some sens for vundle. Hm that actually
might be an ok idea because then I could just run :PluginInstall after
that w/o having to install vundle "manually".