Refactor .vim setup
The idea is to add Vundle as a submodule but not track the other Plugins
that may be installed by Vundle. Instead those will be automatically
installed in the setup.sh script. I cannot be arsed to keep track of the
submodules for all the plugins for vim so I think this setup should be
quite smooth.