Set up git mergetool nvim
It opens up a three-way merge nvim configuration with four windows, like
so:

+-----------+------------+------------+
|           |            |            |
|           |            |            |
|   LOCAL   |    BASE    |   REMOTE   |
+-----------+------------+------------+
|                                     |
|                                     |
|             (edit me)               |
+-------------------------------------+

The LOCAL one is the version from "this" branch, the REMOTE is the one
your merging in, the BASE is how it looked like prior to any side doing
any work. At the bottom you have the resulting merged file.
To pick a change from e.g. LOCAL do `:diffget LOCAL` and so on.
TO conclude the merge `:xa` (equivalent to `:wqa` (write-quit-all)).