🐛 Don't autoUpgrade
It explains my issues with "hm-session-vars" every now and then - this
option runs nixos-rebuild switch every night, and was by default using
/etc/nixos/configuration.nix!
It is kind of neat and it is configurable:
https://search.nixos.org/options?channel=unstable&show=system.autoUpgrade.flake&from=0&size=50&sort=relevance&type=packages&query=system.autoupgrade
so that I could point it at my flake and automaticlly update the nixpkgs
input etc. But, then I figured, do I want to do that?
For the time being, no I believe not.