🐛 Correctly disable markdown in tree-sitter
Previous attempt at disabling it was impotent.
I didn't explicitly ever set it as installed, but unfortunately it
comes bundled with neovim in nixpkgs it seems. This leads to a number
of issues as it appears to conflict syntax highlighting groups with
preservim/vim-markdown.
I did a quick test - either go with vim-markdown or with treesitter
markdown. treesitter looks *perhaps* a little nicer with e.g. code
blocks. But doesn't support latex, or folding.
Though I don't use folding that much so whatever on that part.
Conceptually I prefer TS and that could make me get rid of one plugin,
but latex folding is quite nice (though again uncommon I actually use it
be fair). For now I'm keeping this but might change in the future.